[batch] Use Timings with no consideration of job deletion
